Add draws from the posterior predictive of a rethinking model to a data frame

Adds draws from the posterior predictive distribution of a rethinking model to a data frame. Provides support for tidybayes::predicted_draws() / tidybayes::add_predicted_draws() for models from the rethinking package.

# S3 method for ulam
predicted_draws(
  model,
  newdata,
  prediction = ".prediction",
  ...,
  n = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)

# S3 method for quap
predicted_draws(
  model,
  newdata,
  prediction = ".prediction",
  ...,
  n = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)

# S3 method for map
predicted_draws(
  model,
  newdata,
  prediction = ".prediction",
  ...,
  n = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)

# S3 method for map2stan
predicted_draws(
  model,
  newdata,
  prediction = ".prediction",
  ...,
  n = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)

Arguments

model
A model fit using rethinking::quap(), rethinking::ulam(), rethinking::map(), or rethinking::map2stan().

newdata
Data frame to generate predictions from. If omitted, most model types will generate predictions from the data used to fit the model.

prediction
The name of the output column for predicted_draws; default ".prediction".

...
Optional parameters passed on to rethinking::sim(). The most pertinent are:

post: Optional samples from posterior. If missing, simulates samples using n.

n
The number of draws per prediction to return. When NULL (the default), rethinking::ulam() and rethinking::map2stan() models return all draws; rethinking::quap() and rethinking::map() models return 5000 draws.

seed
A seed to use when subsampling draws (i.e. when n is not NULL).

re_formula, category
Not used with this model type.